// src/lib/types/contentful.d.ts
// THIS FILE IS AUTOMATICALLY GENERATED. DO NOT MODIFY IT.

import { Asset, Entry } from 'contentful';
import { Document } from '@contentful/rich-text-types';

export interface IProjectFields {
  /** slug */
  slug: string;

  /** Primary Colour */
  primaryColour: string;

  /** Secondary Colour */
  secondaryColour: string;

  /** Display Order */
  displayOrder: number;

  /** Title Display */
  titleDisplay: string;

  /** Description */
  description: string;

  /** Name */
  name: string;

  /** Tagline */
  tagline: string;

  /** Key Roles */
  keyRoles: string[];

  /** url */
  url?: string | undefined;

  /** Categories */
  categories: string[];

  /** TechTitle */
  techTitle: string;

  /** TechDescription */
  techDescription: string;

  /** Tech */
  tech: string[];

  /** Main Image */
  mainImage: Asset;

  /** Mashup Image */
  mashupImage: Asset;

  /** Mobile Images */
  mobileImages?: Asset[] | undefined;

  /** Display Image */
  displayImage: Asset;

  /** Video Fallback */
  videoFallback?: Asset | undefined;

  /** Video */
  video?: Asset | undefined;
}

/** A body of development and design work. */

export interface IProject extends Entry<IProjectFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: 'project';
        linkType: 'ContentType';
        type: 'Link';
      };
    };
  };
}

export type CONTENT_TYPE = 'project';

export type LOCALE_CODE = 'en-GB' | 'en-US';

export type CONTENTFUL_DEFAULT_LOCALE_CODE = 'en-US';